chore(docker): Include sources and Tests in image
We intend to use this image as a universal E2E Test image. Those tests
are defined in Java JUnit and we intend to run them via `mvn -p
e2e-tests test` from inside the container.

The idea is, that this container can be used to test any X1/X2/X3
conformant backend, while it also ships with a (wire)mock Backend as
default.

Co-authored-by: Jannik Volkland <volkland@sipgate.de>
Co-authored-by: Sven Uhlig <uhlig@sipgate.de>