test(x2): add test that X2 could be received by the Simulator.
To test that we added a REST trigger in the Simulator ("/sip") that receives SIP. It is then encoded to X2 and sent with an x2x3Client to its own X2 interface.

To check if the correct X2 has been received we store
the received X2 in then X2X3Memory component, which we then query with an also new REST call "/x2/last".

Co-authored-by: Sven Uhlig <uhlig@sipgate.de>